+++ examples/Carla_Examples/Carla-Demo-Script-Birdview.py
@@ -2,7 +2,6 @@ import pygame
 import os
 import sys
 import numpy as np
-from PIL import Image as PImage
 import cv2
 import imageio
 from dotenv import load_dotenv
