+++ examples/Demo_Drive.py
@@ -4,11 +4,10 @@ import sys
 from PIL import Image as PImage
 import imageio
 import numpy as np
-import cv2
 from tqdm import tqdm
 import argparse
 
-os.environ["IAI_MOCK_API"] = "0"
+os.environ["IAI_MOCK_API"] = "1"
 os.environ["IAI_DEV"] = "1"
 # os.environ["IAI_DEV_URL"] = "http://localhost:8888"
 
@@ -31,18 +30,16 @@ response = iai.location_info(location=args.location)
 file_name = args.location.replace(":", "_")
 if response.osm_map is not None:
     file_path = f"{file_name}.osm"
-    with open(file_path, "w") as f:
-        f.write(response.osm_map[0])
+    response.osm_map.save_osm_file(file_path)
 if response.birdview_image is not None:
     file_path = f"{file_name}.jpg"
-    rendered_map = np.array(response.birdview_image, dtype=np.uint8)
-    image = cv2.imdecode(rendered_map, cv2.IMREAD_COLOR)
-    cv2.imwrite(file_path, image)
+    response.birdview_image.decode_and_save(file_path)
 simulation = iai.BasicCosimulation(
     location=args.location,
     agent_count=10,
     monitor_infractions=True,
     render_birdview=True,
+    ego_agent_mask=[False] * 10,
 )
 frames = []
 pbar = tqdm(range(50))
@@ -55,8 +52,7 @@ for i in pbar:
         + f"Wrong-way rate: {100*np.array(wrong_way).mean():.2f}%"
     )
 
-    birdview = np.array(simulation.birdview, dtype=np.uint8)
-    image = cv2.imdecode(birdview, cv2.IMREAD_COLOR)
+    image = simulation.birdview.decode()
     frames.append(image)
     im = PImage.fromarray(image)
 imageio.mimsave("iai-drive.gif", np.array(frames), format="GIF-PIL")
