+++ invertedai/utils.py
@@ -4,7 +4,9 @@ import re
 from typing import Dict, Optional
 from requests.auth import AuthBase
 import invertedai as iai
-from invertedai import error, api_resources
+import invertedai.api
+import invertedai.api.config
+from invertedai import error, api
 import logging
 
 TIMEOUT_SECS = 600
@@ -34,7 +36,7 @@ class Session:
         self.session.auth = APITokenAuth(api_token)
 
     def use_mock_api(self, use_mock: bool = True) -> None:
-        api_resources.mock_api = use_mock
+        invertedai.api.config.mock_api = use_mock
         if use_mock:
             iai.logger.warning(
                 'Using mock Inverted AI API - predictions will be trivial'
